Extracted argument parsing to mod
